pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-sql-agent"
version = "1.0.0"
description = "Production-grade multi-agent RAG system for natural language to SQL"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Matteo Vitellaro", email = "mattrovit@gmail.com"},
]
keywords = ["rag", "sql", "multi-agent", "langchain", "faiss", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "litellm>=1.40.0",
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "structlog>=24.1.0",
    "rich>=13.7.0",
    "faker>=25.0.0",
    "streamlit>=1.35.0",
    "watchdog>=4.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
rag-sql-seed = "scripts.seed_db:main"
rag-sql-index = "scripts.build_index:main"
rag-sql-serve = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src", "scripts"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "B", "SIM"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
